refactor(web): move to provider pattern
